Move sync files into wled00/sync/, keep udp.cpp at top level
udp.cpp stays where contributors expect it (still just the
handleNotifications() dispatcher); the six files it dispatches to move
into wled00/sync/ to signal they're a cohesive unit distinct from the
rest of wled00/'s flat file layout.

Only change needed beyond the file moves: udp.cpp's #include "sync.h"
becomes #include "sync/sync.h" since udp.cpp no longer shares a
directory with it. The three moved files that also include sync.h
(sync_notifier.cpp, realtime_udp.cpp, espnow_sync.cpp) need no change,
since they now share a directory with it. Confirmed via the actual
compiler invocation that wled00/ is already passed as an explicit -I,
so #include "wled.h" continues to resolve correctly from the
subfolder with no build_flags/platformio.ini changes required.

Verified: nodemcuv2 builds and links cleanly (933555 bytes flash,
+16 bytes vs. before the move - just the changed include path).
